Added LabJack TM stream and link to config
Made some minor fixes to the LabJack data link class itself. Added calls to updateStats in LabJackDataLink class as well as SerialDataLink class to correctly update the number of packets received by these links. Fixed issue with pin naming in LABJ_XTCE. This is as a result of the testing conducted on 10/10/24.