Reduced packet size, reading DIO_STATE
Reduced labjack packet size by reducing the size of a digital reading from 2 to 1 bit. Also, instead of reading individual digital pins, we now do a singular read of the DIO_STATE register which contains the state of all digital registers. Additionally, it was discovered that it is problematic to read individual digital pins because upon being read, they switch their directionality to input.