Implemented FC commands and acks
Acks are published to commands by creating the expected string of a command's ack as soon as it is sent and placing that expected ack in a map where it acts as the key and maps to the ID of its command. Whenever data is received via the serial port, the data is compared to the existing set of expected acks and if a match is found, the ack is published to the command it maps to.